// src/app/shared/helper.enum.ts
export enum ProfileMenuOption {
    accountSettings = "ACCOUNT SETTINGS",
    profileInformation = "Profile Information",
    manageAddresses = "Manage Addresses",
    panCardInfo = "Pan Card Information",
    logout = "Logout",
    payments = "PAYMENTS",
    phonePeWallet = "PhonePe Wallet",
    giftCards = "Gift Cards",
    payLater = "Pay Later",
    savedCards = "Saved Cards",
    myStuff = "MY STUFF",
    myOrders = "My Orders",
    myRewards = "My Rewards",
    myReviews = "My Reviews & Ratings",
    allNotifications = "All Notifications",
    myWishlist = "My Wishlist"
}